fix: squad header bar responsiveness
